:bug: remove lodash from the edge
