+++ pages/api/sign-grant-agreement.ts
@@ -2,7 +2,7 @@ import { getProjectById } from '@/db/project'
 import { NextRequest, NextResponse } from 'next/server'
 import { createEdgeClient } from './_db'
 import { getBidsByProject } from '@/db/bid'
-import { checkGrantFundingReady } from '@/utils/math'
+import { checkGrantFundingReady } from '@/utils/checks'
 
 export const config = {
   runtime: 'edge',
